# packages/api-client/src/models/CreateTaskRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateTaskRequestDto = {
    /**
     * タスクのタイトル
     */
    title: string;
};