# packages/api-client/src/models/SignInRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SignInRequestDto = {
    email: string;
    password: string;
};