# packages/api-client/src/models/UpdateTaskRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateTaskRequestDto = {
    title?: string;
    isDone?: boolean;
};